web_vla_agent/pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "web-vla-agent"
version = "0.1.0"
description = "Publication-grade Vision-Language-Action Web Agent"
requires-python = ">=3.10"
license = {text = "MIT"}

dependencies = [
    # Core
    "torch>=2.0",
    "torchvision",
    "transformers>=4.36",
    "peft>=0.7",
    "accelerate>=0.25",

    # Graph
    "torch-geometric>=2.4",

    # Multimodal
    "Pillow>=10.0",
    "numpy>=1.24",
    "sentence-transformers>=2.2",

    # Browser
    "playwright>=1.40",

    # Data
    "datasets>=2.14",
    "beautifulsoup4>=4.12",
    "lxml>=4.9",

    # Config & Utils
    "pyyaml>=6.0",
    "tqdm>=4.65",

    # Tracking
    "wandb>=0.16",
]

[project.optional-dependencies]
dev = ["pytest>=7.4", "ruff>=0.1", "pytest-asyncio"]
rl = ["gymnasium>=0.29", "stable-baselines3>=2.1"]

[project.scripts]
vla-train = "training.train_supervised:main"
vla-eval = "evaluation.evaluate:main"
vla-agent = "main:main"

[tool.ruff]
line-length = 100
target-version = "py310"
